requirements.txt
netmiko
paramiko
python-dotenv
pysnmp

# aiofiles==0.8.0
# aiohttp==3.8.1
# aiohttp-swagger==1.0.16
# aiosignal==1.2.0
# aniso8601==9.0.1
# astroid==2.9.3
# asttokens==2.0.5
# async-lru==1.0.2
# async-timeout==4.0.2
# asyncssh==2.9.0
# attrs==21.4.0
# backcall==0.2.0
# bcrypt==3.2.0
# black==21.12b0
# blessings==1.7
# bpython==0.22.1
# certifi==2021.10.8
# cffi==1.15.0
# chardet==4.0.0
# charset-normalizer==2.0.12
# click==7.1.2
# colorama==0.4.4
# commonmark==0.9.1
# cryptography==36.0.1
# curtsies==0.3.10
# cwcwidth==0.1.6
# darglint==1.8.1
# decorator==5.1.1
# dill==0.3.4
# distro==1.7.0
# executing==0.8.3
# fancycompleter==0.9.1
# flake8==4.0.1
# flake8-bugbear==22.1.11
# flake8-docstrings==1.6.0
# flake8-mypy==17.8.0
# Flask==1.1.4
# flask-restx==0.5.1
# frozenlist==1.3.0
# future==0.18.2
# genie==21.9
# genie.libs.clean==21.9.3
# genie.libs.conf==21.9
# genie.libs.filetransferutils==21.9
# genie.libs.health==21.9.1
# genie.libs.ops==21.9
# genie.libs.parser==21.9
# genie.libs.sdk==21.9
# greenlet==1.1.2
# hvac==0.11.2
# idna==3.3
# iniconfig==1.1.1
# ipdb==0.13.9
# ipython==8.1.0
# isort==5.10.1
# itsdangerous==1.1.0
# jedi==0.18.1
# Jinja2==2.11.3
# jsonpickle==2.1.0
# jsonschema==4.4.0
# junit-xml==1.9
# lazy-object-proxy==1.7.1
# lxml==4.6.5
# MarkupSafe==2.0.1
# matplotlib-inline==0.1.3
# mccabe==0.6.1
# multidict==6.0.2
# mypy==0.931
# mypy-extensions==0.4.3
# ncclient==0.6.12
# netaddr==0.8.0
# netmiko==3.4.0
# ntc-templates==3.0.0
# packaging==21.3
# paramiko==2.7.2
# parso==0.8.3
# pathspec==0.9.0
# pdbpp==0.10.3
# pexpect==4.8.0
# pickleshare==0.7.5
# platformdirs==2.5.1
# pluggy==1.0.0
# prettytable==3.1.1
# prompt-toolkit==3.0.28
# psutil==5.9.0
# ptyprocess==0.7.0
# pure-eval==0.2.2
# py==1.11.0
# pyang==2.5.2
# pyats==21.9
# pyats.aereport==21.9
# pyats.aetest==21.9
# pyats.async==21.9
# pyats.connections==21.9
# pyats.datastructures==21.9
# pyats.easypy==21.9
# pyats.kleenex==21.9
# pyats.log==21.9
# pyats.reporter==21.9
# pyats.results==21.9
# pyats.tcl==21.9
# pyats.topology==21.9
# pyats.utils==21.9
# pycodestyle==2.8.0
# pycparser==2.21
# pydocstyle==6.1.1
# pyflakes==2.4.0
# pyftpdlib==1.5.6
# Pygments==2.11.2
# pylama==8.3.7
# pylint==2.12.2
# PyNaCl==1.5.0
# pyparsing==3.0.7
# pyrepl==0.9.0
# pyrsistent==0.18.1
# pyserial==3.5
# pytest==7.0.1
# python-engineio==3.14.2
# python-socketio==4.6.1
# pytz==2021.3
# pyxdg==0.27
# PyYAML==5.4.1
# requests==2.26.0
# rich==11.2.0
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.6
# scp==0.14.4
# scrapli==2021.7.30
# scrapli-netconf==2021.7.30
# six==1.16.0
# snowballstemmer==2.2.0
# ssh2-python==0.27.0
# stack-data==0.2.0
# tenacity==8.0.1
# textfsm==1.1.2
# tftpy==0.8.0
# toml==0.10.2
# tomli==1.2.3
# tqdm==4.63.0
# traitlets==5.1.1
# ttp==0.8.4
# types-click==7.1.8
# types-Flask==1.1.6
# types-Jinja2==2.11.9
# types-MarkupSafe==1.1.10
# types-requests==2.27.11
# types-urllib3==1.26.10
# types-Werkzeug==1.0.9
# typing_extensions==4.1.1
# unicon==21.9
# unicon.plugins==21.9
# urllib3==1.26.8
# wcwidth==0.2.5
# Werkzeug==1.0.1
# wmctrl==0.4
# wrapt==1.13.3
# xmltodict==0.12.0
# yamllint==1.26.3
# yarl==1.7.2
